use _ElementOfUnknownGroup and safe-scalarmult to fix arbitrary_element
Our (faster) multiplication routine is only correct for elements whose
orders are a multiple of 'l'. For the other options (1,2,4,8), the
double-and-add ladder leads to adding equal points to each other (the
doubling eventually gets to the identity Zero, and adding Zero to Zero
gave a bad result).

This splits scalarmult_extended() two functions. The first (faster)
handles the normal case where the point is part of the 1*l subgroup. It
also handles points in the 2l/4l/8l subgroups. And it handles Zero as a
special case.

The second (slower) handles anything, and uses the (slower) unified
addition routine.

To make these easier to access, the new _ElementOfUnknownGroup class
uses the second scalarmult exclusively. And arbitrary_element() works
with an _ElementOfUnknownGroup until it finishes the correct-group
membership check, at which point it promotes the point to an Element.